% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVIF.R
\name{CalculateVIF}
\alias{CalculateVIF}
\title{Calculate Variance inflation factor (VIF) for a given model}
\usage{
CalculateVIF(input_data, sdq, var_list, sign_var_list, conf_list)
}
\arguments{
\item{input_data}{A dataframe with a complete dataset}

\item{sdq}{A string indicating which SDQ score will be plugged into the model}

\item{var_list}{A character vector containing a list of exposures for which regression will be run}

\item{sign_var_list}{A character vector containing a list of exposures detected as significanly associated with SDQ in ExWAS}

\item{conf_list}{A character vector specifying confounders to be added in the regression model}
}
\value{
A character vector with exposures that are not collinear and can be plugged into a multpile regression model
}
\description{
Calculate Variance inflation factor (VIF) for a given model
}
