% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DadosPalma.R
\docType{data}
\name{DadosPalma}
\alias{DadosPalma}
\title{Dados de exemplo de um experimento sem repeticoes.}
\format{
Um data frame com uma variavel por coluna.
}
\usage{
data(DadosPalma)
}
\description{
Exemplo com um conjunto de dados
de um experimento sem delineamento estatistico.Neste caso,deve-se ter uma tabela com
uma variavel por coluna
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK6OUnWVs-hnyVdCB1xuxbVs>
}
\author{
Alcinei Mistico Azevedo, \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
