% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleRegression.R
\name{MultipleRegression}
\alias{MultipleRegression}
\title{Function for running a multiple linear regression model with listed variables}
\usage{
MultipleRegression(input_data, mlr_var_list, variable_list, conf_list, sdq)
}
\arguments{
\item{input_data}{"mids" object (imputed dataset)}

\item{mlr_var_list}{A character vector containing the names of exposures to include in the multiple model}

\item{variable_list}{A dataframe with short and full names of exposure variables}

\item{conf_list}{A character vector specifying names of model adjustment factors (confounders)}

\item{sdq}{A string defining which SDQ score will be fitted}
}
\value{
A dataframe with regression results for all exposures in the multiple model
}
\description{
Function for running a multiple linear regression model with listed variables
}
