% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictModel.R
\name{PredictModel}
\alias{PredictModel}
\title{Funcao para obter os valores preditos pelo modelo de regressao}
\usage{
PredictModel(object,newdata=NULL)
}
\arguments{
\item{object}{Objeto criado pela funcao `FitModel`}

\item{newdata}{Objeto do tipo `data.frame` com os valores das variaveis explicativas
a serem preditas.}
}
\value{
Retorna os valores preditos pelo modelo de regressao.
}
\description{
Esta funcao prediz os valores do modelo ajustado.
}
\examples{
data("DadosPalma")
X=DadosPalma[,-7]
Y=DadosPalma[,7]
model=CreateModel(X,Intercept = TRUE, Simple =TRUE, InteractionLL = TRUE)
model
Ajust=FitModel(X,Y,Model=model)
Ajust
PredictModel(Ajust)

}
\references{
Tutoriais onlines:
https://www.youtube.com/channel/UCDGyvLCJnv9RtTY1YMBMVNQ
}
\seealso{
\code{\link{FitModel}}, \code{\link{CreateModel}}, \code{\link{PredictModel}}
}
